% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List2VennDiagram.R
\name{vennPair}
\alias{vennPair}
\title{2 Set Venn}
\usage{
vennPair(sets, sets_col, ...)
}
\arguments{
\item{sets}{List of sets}

\item{sets_col}{vector mapping set names to colors}

\item{...}{passed through to VennDiagram::draw.pairwise.venn}
}
\description{
Given a list of sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed trhough.
}
\examples{
x=vennPair(list(A=c(1:5),B=c(3:10)),c("A"="red", "B"="blue"))
}
