% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sets2VennDiagram.R
\name{vennSetPair}
\alias{vennSetPair}
\title{2 Set Venn}
\usage{
vennSetPair(sets, ...)
}
\arguments{
\item{sets}{A List of 2 sets, one for each category of the diagram}

\item{...}{passed through to VennDiagram::draw.pairwise.venn}
}
\description{
Given a list of sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed trhough.
}
\examples{
x=vennSetPair(list(A=c(1:5),B=c(3:10)))
}
