% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sets2VennDiagram.R
\name{vennSetQintuple}
\alias{vennSetQintuple}
\title{5 Set Venn (Euler)}
\usage{
vennSetQintuple(sets, ...)
}
\arguments{
\item{sets}{A List of 5 sets, one for each category of the diagram}

\item{...}{passed through to VennDiagram::draw.quintuple.venn}
}
\description{
Given a list of sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed through.
}
\examples{

x=vennSetQintuple(list(A=c(1:5),B=c(3:10),C=c(8:20,1),D=c(1:2,3:4,15:20,99:100),E=(50:100)))
}
