% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List2VennDiagram.R
\name{vennSetTriple}
\alias{vennSetTriple}
\title{3 Set Venn (Euler)}
\usage{
vennSetTriple(sets, sets_col, ...)
}
\arguments{
\item{sets}{List of sets}

\item{sets_col}{vector mapping set names to colors}

\item{...}{passed through to VennDiagram::draw.triple.venn}
}
\description{
Given a list of sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed trhough.
}
\examples{
x=vennSetTriple(list(A=c(1:5),B=c(3:10),C=c(8:20,1)),c("A"="red", "B"="blue", "C"="green"))
}
