% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List2VennDiagram.R
\name{vennSetQintuple}
\alias{vennSetQintuple}
\title{5 Set Venn (Euler)}
\usage{
vennSetQintuple(sets, sets_col, ...)
}
\arguments{
\item{sets}{List of sets}

\item{sets_col}{vector mapping set names to colors}

\item{...}{passed through to VennDiagram::draw.quintuple.venn}
}
\description{
Given a list of sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed through.
}
\examples{

x=vennSetQintuple(list(A=c(1:5),B=c(3:10),C=c(8:20,1),D=c(1:2,3:4,15:20,99:100),E=(50:100)),c("A"="red", "B"="blue", "C"="green", "D"="purple", "E"="beige"))
}
