% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sets2VennDiagram.R
\name{vennSetQuad}
\alias{vennSetQuad}
\title{4 Set Venn (Euler)}
\usage{
vennSetQuad(sets, sets_col, ...)
}
\arguments{
\item{sets}{List of sets}

\item{sets_col}{vector mapping set names to colors (coming soon)}

\item{...}{passed through to VennDiagram::draw.quad.venn}
}
\description{
Given a list of sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed through.
}
\examples{

x=vennSetQuad(list(A=c(1:5),B=c(3:10),C=c(8:20,1),D=c(1:2,3:4,15:20,99:100)))
}
