% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List2VennDiagram.R
\name{vennSets}
\alias{vennSets}
\title{Venn Diagram from a list of sets}
\usage{
vennSets(sets, sets_col, ...)
}
\arguments{
\item{sets}{List of sets (2-5)}

\item{sets_col}{vector mapping set names to colors}

\item{...}{passed through to VennDiagram::draw.*.venn}
}
\description{
Given a list of 2 to 5 sets, draw the Venn Diagram depecting their overlap.
All parameters for VennDiagram will be passed trhough.
}
\examples{
x=vennSets(list(A=c(1:5),B=c(3:10)),c("A"="red", "B"="blue"))
y=vennSets(list(A=c(1:5),B=c(3:10),C=c(8:20,1)),c("A"="red", "B"="blue", "C"="green"))
}
